Multiplying 2 Amounts results in an Amount, dividing Amount by Amount results in Amount. Use specific conversion functions by casting to Amount or using .toInt() to convert between Amount and int rather than implicit conversions via operators

+++ src/primitives/transaction.cpp
@@ -30,7 +30,7 @@ std::string CTxIn::ToString() const {
 }
 
 std::string CTxOut::ToString() const {
-    return strprintf("CTxOut(nValue=%d.%08d, scriptPubKey=%s)", nValue / COIN,
+    return strprintf("CTxOut(nValue=%d.%08d, scriptPubKey=%s)", (nValue / COIN).toInt(),
                      (nValue % COIN).toInt(),
                      HexStr(scriptPubKey).substr(0, 30));
 }
